Update setup.exe to setupodt.exe
The executable file for the Office Deployment Tool is now named setupodt.exe, instead of setup.exe. This change took effect with Version 16.0.13328.20356 (setupodt.exe version 16.0.13328.20336), which was released on November 10, 2020. We recommend that you always download and use the most current version of the Office Deployment Tool. Therefore, if you have scripts or batch files that refer to setup.exe, you need to update them to refer to setupodt.exe.

Allowed values:

    Channel="BetaChannel"
    Channel="CurrentPreview"
    Channel="Current"
    Channel="MonthlyEnterprise"
    Channel="SemiAnnualPreview"
    Channel="SemiAnnual"